+++ python/src/models/deep/nstransformer.py
@@ -3,15 +3,14 @@ Non-stationary Transformer for Time Series Forecasting.
 Adapted from the Time-Series-Library.
 """
 import torch
-from torch import nn
-
-from .modules import DataEmbedding
-from .modules import (
+from torch import nn 
+from python.src.models.deep.modules import (
     AttentionLayer,
     DSAttention,
     Normalization,
     SkipConnection,
     Transpose,
+    DataEmbedding
 )