+++ python/src/models/deep/tsmamba.py
@@ -5,7 +5,7 @@ Time Series Mamba implementation.
 import torch
 from torch import nn
 
-from .modules.mamba_block import MambaBlock
+from python.src.models.deep.modules.mamba_block import MambaBlock
 
 
 class TSMamba(nn.Module):